pyproject.toml
[project]
name = "helphand"
version = "0.1.0"
description = ""
authors = [
    {name = "Diamond Land"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram (>=3.21.0,<4.0.0)",
    "loguru (==0.7.2)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "aerich (>=0.9.1,<0.10.0)",
    "tomli (>=2.2.1,<3.0.0)",
    "tomlkit (>=0.13.3,<0.14.0)",
    "psutil (>=7.0.0,<8.0.0)"
]


[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.scripts]
start = "src.main:main"



[tool.aerich]
tortoise_orm = "src.config.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
